+++ src/test/ExSln2/LedBlinker/test_stuff/ExTypeOverride.cs
@@ -29,4 +29,9 @@ public class ExTypeOverride : FinObj
     {
         this.pin = pin;
     }
+
+    public void set_pin_use_example()
+    {
+        set_pin(FinC.EchoToC("GPIO_PIN_0"));
+    }
 }