+++ src/test/ExSln2/c99/gen/hal_ExTypeOverride.h
@@ -28,3 +28,5 @@ struct hal_ExTypeOverride
 void hal_ExTypeOverride_ctor(hal_ExTypeOverride * self, GPIO_TypeDef * port, uint16_t pin);
 
 void hal_ExTypeOverride_set_pin(hal_ExTypeOverride * self, uint16_t pin);
+
+void hal_ExTypeOverride_set_pin_use_example(hal_ExTypeOverride * self);